ensure idempotency for post request

+++ models/recipes.models.js
@@ -1,4 +1,3 @@
-const { json } = require('express');
 const {readFile, writeFile} = require('fs/promises');
 
 exports.fetchRecipes = async (query) => {
@@ -6,7 +5,9 @@ exports.fetchRecipes = async (query) => {
     const parsedRecipes = JSON.parse(recipes);
 
     if (query?.exclude_ingredients) {
+        //list of unwanted ingredients
         const excludedIngredientsArr = query.exclude_ingredients.split(',')
+        // format query ingredients for to account for plural and singular spelling
         excludedIngredientsArr.forEach((ingredient)=>{
             if (ingredient[ingredient.length-1] === 's') {
                 const ingredientSRemoved = ingredient.slice(0, -1)
@@ -21,6 +22,7 @@ exports.fetchRecipes = async (query) => {
             }
         })
 
+        //filter out unwanted recipes
         const filteredRecipes = parsedRecipes.filter((recipe)=>{
            return !recipe.ingredients.some(({name})=>{
                return excludedIngredientsArr.includes(name)
@@ -39,21 +41,30 @@ exports.fetchRecipeById = async (id) => {
     return recipe[0]
 }
 
+
 exports.addRecipe = async (recipe) => {
     const recipes = await this.fetchRecipes();
 
+    //check recipe already exists
+    let matchingRecipeIndex = ''
+    const doesRecipeExist =  recipes.some((dataRecipe, index)=> {
+        matchingRecipeIndex = index
+        return JSON.stringify(dataRecipe.ingredients) === JSON.stringify(recipe.ingredients)
+    })
+    if (doesRecipeExist === true) {
+       return recipes[matchingRecipeIndex].id;
+    }
+
+    // find largest id number to generate sequential id
     const recipeIds = recipes.map((recipe)=>{
         return recipe.id.match(/\d+$/)[0]
     })
-
     const largestId = Math.max(...recipeIds)
-
     recipe.id = `recipe-${largestId+1}`
 
+    // write over data with replication plus new recipe
     const recipesAddRecipe = [...recipes, recipe]
-
     await writeFile('./data/data.json', JSON.stringify(recipesAddRecipe, null, 2))
 
     return recipe.id
-    
 }
